refactor(SPEC-CLIFIX-HYGIENE-001): M5 update.go decomposition (mechanical move)
Extract two cohesive clusters out of internal/cli/update.go to satisfy the
1,200-line ceiling (REQ-HYG-001-001):

- update_template_sync.go (590 lines): runTemplateSync,
  runTemplateSyncWithReporter, runTemplateSyncWithProgress, toPreviewInputs,
  confirmViaPreview
- update_wizard.go (352 lines): runInitWizard, applyWizardConfig

update.go 1905 -> 1005 lines. Mechanical move only: function bodies are
byte-identical to the pre-split ranges, the diff on update.go is pure
deletion, and only the import blocks were adjusted (goimports). Same
package, no renames, no export changes, no logic edits.

Characterization suite green pre- and post-split.

🗿 MoAI
