+++ internal/cli/update.go
@@ -16,20 +16,10 @@ import (
 
 	"github.com/charmbracelet/huh"
 	"github.com/mattn/go-isatty"
-	"github.com/modu-ai/moai-adk/internal/cli/printer"
-	"github.com/modu-ai/moai-adk/internal/cli/uikit"
-	"github.com/modu-ai/moai-adk/internal/cli/update"
-	"github.com/modu-ai/moai-adk/internal/cli/update/backup"
 	"github.com/modu-ai/moai-adk/internal/cli/update/deploy"
-	updatemerge "github.com/modu-ai/moai-adk/internal/cli/update/merge"
-	"github.com/modu-ai/moai-adk/internal/cli/update/plan"
 	"github.com/modu-ai/moai-adk/internal/cli/update/report"
-	"github.com/modu-ai/moai-adk/internal/cli/wizard"
 	"github.com/modu-ai/moai-adk/internal/config"
-	"github.com/modu-ai/moai-adk/internal/core/project"
 	"github.com/modu-ai/moai-adk/internal/defs"
-	"github.com/modu-ai/moai-adk/internal/manifest"
-	"github.com/modu-ai/moai-adk/internal/merge"
 	"github.com/modu-ai/moai-adk/internal/profile"
 	"github.com/modu-ai/moai-adk/internal/runtime/gobin"
 	"github.com/modu-ai/moai-adk/internal/shell"
@@ -613,565 +603,6 @@ func reexecNewBinary() error {
 	return syscall.Exec(exe, os.Args, os.Environ())
 }
 
-// runTemplateSync synchronizes embedded templates with the project directory.
-// It performs a quick version comparison first - if the project's template version
-// matches the package version, the sync is skipped for performance (70-80% faster).
-//
-// Template deployment uses a 3-way merge strategy to preserve local modifications.
-// Users are prompted to confirm the merge before proceeding.
-func runTemplateSync(cmd *cobra.Command) error {
-	return runTemplateSyncWithReporter(cmd, nil, false)
-}
-
-// @MX:NOTE: [AUTO] runTemplateSyncWithReporter — M4-S4d-2 DDD migration. Top-level header/section/
-// final-outcome lines are converted to tui.KV / tui.Section / tui.CheckLine / tui.Pill. Sub-step
-// micro messages (\r-prefixed sym* helpers) are preserved because they drive the progress display.
-// Design source: screens.jsx ScreenUpdate.
-//
-// runTemplateSyncWithReporter synchronizes templates with progress reporting.
-func runTemplateSyncWithReporter(cmd *cobra.Command, reporter project.ProgressReporter, skipConfirm bool) error {
-	out := cmd.OutOrStdout()
-	th := resolveTheme()
-	ctx := cmd.Context()
-
-	// Get flags for template sync
-	forceBackup := getBoolFlag(cmd, "force")
-	autoConfirm := getBoolFlag(cmd, "yes")
-
-	// Use current directory as project root
-	projectRoot := "."
-
-	currentVersion := version.GetVersion()
-	// Identity header band (REQ-TUXIU-015): "◆ MoAI-ADK <version> <go-runtime>
-	// · claude" with the version as a solid brand pill.
-	_, _ = fmt.Fprintln(out, renderIdentityBand(currentVersion, th))
-	_, _ = fmt.Fprintln(out, tui.KV("Current version", "moai-adk "+currentVersion, tui.KVOpts{Theme: &th, KeyWidth: 16}))
-	_, _ = fmt.Fprintln(out, tui.CheckLine("run", "Syncing templates", "from embedded filesystem", "", &th))
-
-	if reporter != nil {
-		reporter.StepStart("Version Check", "Checking template version...")
-	}
-
-	// Stage 2: Config Version Comparison (before template sync)
-	// Compare package template_version with project config template_version
-	// If versions match, skip sync for performance (70-80% faster)
-	packageVersion := version.GetVersion()
-	projectVersion, err := plan.GetProjectConfigVersion(projectRoot)
-	if err == nil && packageVersion == projectVersion && !forceBackup {
-		if reporter != nil {
-			reporter.StepComplete("Already up-to-date")
-		}
-		_, _ = fmt.Fprintln(out)
-		_, _ = fmt.Fprintln(out, tui.Pill(tui.PillOpts{Kind: tui.PillOk, Solid: false, Label: report.RenderOutcome(report.OutcomeAlreadyUpToDate, 0, ""), Theme: &th}))
-		return nil
-	}
-
-	if reporter != nil {
-		reporter.StepComplete("Version check complete")
-	}
-
-	if reporter != nil {
-		reporter.StepStart("Loading Templates", "Reading embedded templates...")
-	}
-
-	// Load embedded templates
-	embedded, err := template.EmbeddedTemplates()
-	if err != nil {
-		if reporter != nil {
-			reporter.StepError(err)
-		}
-		return fmt.Errorf("load embedded templates: %w", err)
-	}
-
-	if reporter != nil {
-		reporter.StepComplete("Templates loaded")
-	}
-
-	if reporter != nil {
-		reporter.StepStart("Loading Manifest", "Reading project manifest...")
-	}
-
-	// Initialize manifest manager
-	mgr := manifest.NewManager()
-	if _, err := mgr.Load(projectRoot); err != nil {
-		if reporter != nil {
-			reporter.StepError(err)
-		}
-		return fmt.Errorf("load manifest: %w", err)
-	}
-
-	if reporter != nil {
-		reporter.StepComplete("Manifest loaded")
-	}
-
-	// Create renderer for template variable substitution
-	renderer := template.NewRenderer(embedded)
-
-	// Create deployer with renderer and force update enabled for template sync
-	// This ensures template files are rendered (.tmpl -> actual file) and updated even if they exist
-	deployer := template.NewDeployerWithRendererAndForceUpdate(embedded, renderer, true)
-
-	// Analyze merge changes. The "Analyzing merge changes" header + the
-	// classification card are shown only when this function owns the
-	// confirmation flow (skipConfirm=false). When skipConfirm=true the caller
-	// (runTemplateSyncWithProgress) already printed both before the user's
-	// y/n confirmation — reprinting them here duplicates the visible output.
-	analysis := updatemerge.AnalyzeMergeChanges(deployer, projectRoot)
-
-	if !skipConfirm {
-		_, _ = fmt.Fprintln(out)
-		_, _ = fmt.Fprintln(out, tui.Section("Analyzing merge changes", tui.SectionOpts{Theme: &th}))
-		// Card-style classification summary (REQ-TUXIU-010/011): accent box with
-		// up to three count pills; zero-count pills omitted; suppressed entirely
-		// when the run is clean (all counts zero).
-		addCount, updateCount, conflictCount := classifyUpdateCounts(analysis.Files)
-		if card := renderClassificationSummary(addCount, updateCount, conflictCount, th); card != "" {
-			_, _ = fmt.Fprintln(out, card)
-		}
-	}
-
-	if reporter != nil {
-		reporter.StepUpdate("Found " + fmt.Sprintf("%d files to sync", len(analysis.Files)))
-	}
-
-	// Skip confirmation if --yes flag is provided (CI/CD mode) or pre-confirmed
-	var proceed bool
-	if skipConfirm {
-		proceed = true
-	} else if autoConfirm {
-		proceed = true
-		_, _ = fmt.Fprintln(out, tui.CheckLine("info", "Auto-confirm", "CI/CD mode", "", &th))
-	} else {
-		var err error
-		proceed, err = confirmViaPreview(analysis, projectRoot)
-		if err != nil {
-			if reporter != nil {
-				reporter.StepError(err)
-			}
-			return fmt.Errorf("confirm merge for %d files (risk: %s): %w",
-				len(analysis.Files), analysis.RiskLevel, err)
-		}
-	}
-
-	if !proceed {
-		_, _ = fmt.Fprintln(out)
-		_, _ = fmt.Fprintln(out, tui.Pill(tui.PillOpts{Kind: tui.PillNeutral, Solid: false, Label: "Merge cancelled by user", Theme: &th}))
-		if reporter != nil {
-			reporter.StepError(errors.New("cancelled by user"))
-		}
-		return nil
-	}
-
-	// Deploy templates
-	_, _ = fmt.Fprintln(out)
-	_, _ = fmt.Fprintln(out, tui.Section("Proceeding with template deployment", tui.SectionOpts{Theme: &th}))
-	_, _ = fmt.Fprintln(out)
-
-	// Define deployment steps
-	steps := []struct {
-		name    string
-		message string
-		execute func() error
-	}{
-		{
-			name:    "Backup",
-			message: "Backing up configuration",
-			execute: func() error {
-				// Always backup before update (even with --force)
-				// --force only skips version check, not backup/merge
-				// SPEC-V3R6-UPDATE-PROGRESS-001 M1: tui.ProgressLine replaces
-				// the legacy CR-plus-format pair (REQ-UPR-004).
-				pl := tui.ProgressLine(out, "Backing up .moai/config...", nil)
-				configBackupPath, backupErr := backup.BackupMoaiConfig(projectRoot)
-				if backupErr != nil {
-					pl.Fail(fmt.Sprintf("Backup failed: %v", backupErr))
-					return backupErr
-				}
-				if configBackupPath != "" {
-					pl.Done(".moai/config backed up")
-				} else {
-					pl.Done("No config to backup")
-				}
-				return nil
-			},
-		},
-		{
-			name:    "Validate Templates",
-			message: "Validating all templates before deployment",
-			execute: func() error {
-				homeDir, _ := userHomeDir()
-				goBinPath := detectGoBinPathForUpdate(homeDir)
-				tmplCtx := template.NewTemplateContext(
-					template.WithGoBinPath(goBinPath),
-					template.WithResolvedMoaiPath(resolveMoaiExecutable()),
-					template.WithHomeDir(homeDir),
-					template.WithSmartPATH(template.BuildSmartPATH()),
-					template.WithPlatform(runtime.GOOS),
-					template.WithVersion(version.GetVersion()),
-					template.WithHookOptIn(readHookOptInEnabled(projectRoot)),
-				)
-
-				// SPEC-V3R6-UPDATE-PROGRESS-001 M1: tui.ProgressLine replaces
-				// the legacy CR-plus-format pair (REQ-UPR-004).
-				pl := tui.ProgressLine(out, "Validating templates...", nil)
-				if validateErr := deployer.ValidateAll(ctx, tmplCtx); validateErr != nil {
-					pl.Fail(fmt.Sprintf("Template validation failed: %v", validateErr))
-					return fmt.Errorf("template validation: %w", validateErr)
-				}
-				pl.Done("All templates validated")
-				return nil
-			},
-		},
-		{
-			name:    "Clean Managed Paths",
-			message: "Removing old MoAI-managed files",
-			execute: func() error {
-				return deploy.CleanMoaiManagedPaths(projectRoot, out)
-			},
-		},
-		{
-			name:    "Deploy Templates",
-			message: "Deploying template files",
-			execute: func() error {
-				// SPEC-V3R6-UPDATE-PROGRESS-001 M1: tui.ProgressLine replaces
-				// the legacy CR-plus-format pair (REQ-UPR-004).
-				pl := tui.ProgressLine(out, "Deploying templates...", nil)
-
-				// Build TemplateContext with detected paths for template rendering
-				homeDir, _ := userHomeDir()
-				goBinPath := detectGoBinPathForUpdate(homeDir)
-				tmplCtx := template.NewTemplateContext(
-					template.WithGoBinPath(goBinPath),
-					template.WithResolvedMoaiPath(resolveMoaiExecutable()),
-					template.WithHomeDir(homeDir),
-					template.WithSmartPATH(template.BuildSmartPATH()),
-					template.WithPlatform(runtime.GOOS),
-					template.WithVersion(version.GetVersion()),
-					template.WithHookOptIn(readHookOptInEnabled(projectRoot)),
-				)
-
-				if deployErr := deployer.Deploy(ctx, projectRoot, mgr, tmplCtx); deployErr != nil {
-					pl.Fail(fmt.Sprintf("Deployment failed: %v", deployErr))
-					return fmt.Errorf("deploy templates: %w", deployErr)
-				}
-				pl.Done("Templates deployed")
-				return nil
-			},
-		},
-		{
-			name:    "Restore Settings",
-			message: "Restoring user settings",
-			execute: func() error {
-				// This step's status is tracked via configBackupPath variable
-				// We'll handle this in the main flow
-				return nil
-			},
-		},
-	}
-
-	// Track config backup path for restore step
-	var configBackupPath string
-	// Backup of user's .gitignore content for EntryMerge after deploy
-	var gitignoreBackup []byte
-	// Backups of mergeable files for 3-way merge after deploy
-	var mergeableBackups []updatemerge.FileBackup
-
-	// collectMergeableFiles returns a list of files that should be merged
-	// using the 3-way merge engine during update.
-	// Note: .moai/config/sections/*.yaml files are already handled by
-	// restoreMoaiConfig with 3-way merge, so they are excluded here.
-	collectMergeableFiles := func(projectRoot string) []string {
-		// Fixed mergeable files at project root that are NOT handled by restoreMoaiConfig.
-		// .mcp.json is intentionally absent: MoAI no longer ships an MCP template
-		// (full MCP removal), so a user's .mcp.json is not a merge target.
-		return []string{
-			".claude/settings.json",
-			".moai/status_line.sh",
-		}
-	}
-
-	// Execute each step with progress reporting. The per-step in-flight line is
-	// rendered by the inline tui.ProgressLine primitive inside each step body;
-	// the coarse reporter Step wrapper is intentionally NOT driven here — doing
-	// so double-rendered every step on the stderr channel (the printer.Step
-	// in-flight line) alongside the stdout ProgressLine, producing the two-part
-	// "○…○" spinner residue on a TTY (REQ-TUXIU-020/021). Errors still surface
-	// via reporter.StepError (orphan-safe) below.
-	for i, step := range steps {
-		// Special handling for backup/restore steps; default executes normally
-		switch step.name {
-		case "Backup":
-			// SPEC-V3R6-UPDATE-PROGRESS-001 M1: tui.ProgressLine replaces
-			// the legacy CR-plus-format pair (REQ-UPR-004).
-			plBackup := tui.ProgressLine(out, "Backing up .moai/config...", nil)
-			var backupErr error
-			configBackupPath, backupErr = backup.BackupMoaiConfig(projectRoot)
-			if backupErr != nil {
-				plBackup.Fail(fmt.Sprintf("Backup failed: %v", backupErr))
-				if reporter != nil {
-					reporter.StepError(backupErr)
-				}
-				return backupErr
-			}
-			if configBackupPath != "" {
-				plBackup.Done(".moai/config backed up")
-			} else {
-				plBackup.Done("No config to backup")
-			}
-
-			// SPEC-V3R6-UPDATE-NAMESPACE-PROTECT-001 M3: user-owned namespace backup
-			// (REQ-UNP-004). Sequential after .moai/config backup. Skips silently
-			// when no user-owned content exists (EC-UNP-001).
-			plNsBackup := tui.ProgressLine(out, "Backing up user-owned namespace...", nil)
-			nsBackupPath, nsBackupErr := backupUserOwnedNamespace(projectRoot)
-			if nsBackupErr != nil {
-				plNsBackup.Fail(fmt.Sprintf("Namespace backup failed: %v", nsBackupErr))
-				if reporter != nil {
-					reporter.StepError(nsBackupErr)
-				}
-				return nsBackupErr
-			}
-			if nsBackupPath != "" {
-				// Derive display-friendly project-root-relative path
-				displayNs := nsBackupPath
-				if rel, relErr := filepath.Rel(projectRoot, nsBackupPath); relErr == nil {
-					displayNs = rel
-				}
-				plNsBackup.Done(fmt.Sprintf("User-owned namespace backed up: %s", displayNs))
-			} else {
-				plNsBackup.Done("No user-owned namespace to back up")
-			}
-
-			// SPEC-INTERNAL-SECURITY-001 REQ-SEC-006 (AC-SEC-006a): wire the
-			// pre-modification abort sentinel as a REAL gate on the moai update
-			// deploy path. verifyNamespaceBackupCoverage runs AFTER the backup
-			// completes and BEFORE any destructive deploy step; it aborts with
-			// UPDATE_USER_NAMESPACE_VIOLATION when a user-owned namespace file
-			// on disk would be overwritten without a backup. Normal updates
-			// (no user-owned content, or fully backed up) pass through
-			// unchanged (NFR-SEC-003).
-			if covErr := verifyNamespaceBackupCoverage(projectRoot, nsBackupPath); covErr != nil {
-				plNsBackup.Fail(fmt.Sprintf("Namespace safety check failed: %v", covErr))
-				if reporter != nil {
-					reporter.StepError(covErr)
-				}
-				return covErr
-			}
-
-			// Also backup .gitignore for EntryMerge after deploy
-			gitignorePath := filepath.Join(projectRoot, ".gitignore")
-			if data, readErr := os.ReadFile(gitignorePath); readErr == nil {
-				gitignoreBackup = data
-			}
-			// Backup mergeable files for 3-way merge after deploy
-			mergeableFiles := collectMergeableFiles(projectRoot)
-			for _, mf := range mergeableFiles {
-				mfPath := filepath.Join(projectRoot, mf)
-				if data, readErr := os.ReadFile(mfPath); readErr == nil {
-					mergeableBackups = append(mergeableBackups, updatemerge.FileBackup{Path: mf, Data: data})
-				}
-			}
-		case "Restore Settings":
-			// Handle restore step with captured backup path
-			if configBackupPath != "" {
-				// SPEC-V3R6-UPDATE-PROGRESS-001 M1: tui.ProgressLine replaces
-				// the legacy CR-plus-format pair (REQ-UPR-004).
-				plRestore := tui.ProgressLine(out, "Restoring user settings...", nil)
-				if restoreErr := backup.RestoreMoaiConfig(projectRoot, configBackupPath, func(pr, relPath string, success bool, errOut io.Writer) {
-					// Bridge to the noise-suppression ledger (recordMergeFallback +
-					// updateVerboseMode), which stays in package cli. The closure
-					// captures updateVerboseMode so the backup subpackage does not
-					// need a cross-package mutable-state seam.
-					recordMergeFallback(pr, relPath, success, updateVerboseMode, errOut)
-				}); restoreErr != nil {
-					plRestore.Fail(fmt.Sprintf("Restore failed: %v", restoreErr))
-					if reporter != nil {
-						reporter.StepError(restoreErr)
-					}
-					return restoreErr
-				}
-				plRestore.Done("User settings restored")
-				deletedCount := backup.CleanupOldBackups(projectRoot, 5)
-				if deletedCount > 0 {
-					_, _ = fmt.Fprintf(out, "  %s Cleaned up %d old backup(s)\n", uikit.SymSuccess(), deletedCount)
-				}
-			}
-			// Merge .gitignore: preserve user-added patterns via EntryMerge
-			if len(gitignoreBackup) > 0 {
-				gitignorePath := filepath.Join(projectRoot, ".gitignore")
-				if mergeErr := updatemerge.MergeGitignoreFile(gitignorePath, gitignoreBackup); mergeErr != nil {
-					_, _ = fmt.Fprintf(out, "  %s .gitignore merge warning: %v\n", uikit.SymWarning(), mergeErr)
-				} else {
-					_, _ = fmt.Fprintf(out, "  %s .gitignore user patterns preserved\n", uikit.SymSuccess())
-				}
-			}
-			// Merge user-customized files using 3-way merge engine
-			if len(mergeableBackups) > 0 {
-				if err := updatemerge.MergeUserFiles(projectRoot, mergeableBackups, out); err != nil {
-					_, _ = fmt.Fprintf(out, "  %s File merge warning: %v\n", uikit.SymWarning(), err)
-				}
-			}
-		default:
-			// Execute normal step
-			if err := step.execute(); err != nil {
-				if reporter != nil {
-					reporter.StepError(err)
-				}
-				return err
-			}
-		}
-
-		// Block progress bar reflecting completed/total deploy steps
-		// (REQ-TUXIU-014). Replaces the legacy "N/M steps complete" reporter
-		// text; the bar rides the same stdout channel as the step ProgressLines.
-		_, _ = fmt.Fprintln(out, renderDeployProgress(i+1, len(steps), th))
-	}
-
-	_, _ = fmt.Fprintln(out)
-	// Outcome banner (REQ-TUXIU-016): solid success pill + dim detail note.
-	renderUpdateOutcome(out, len(analysis.Files), configBackupPath, th)
-	report.EmitHooksReviewGuidance(out)
-
-	_, _ = fmt.Fprintln(out)
-	_, _ = fmt.Fprintln(out, "To reconfigure project settings (development mode, git, model policy), run:")
-	_, _ = fmt.Fprintln(out, "   moai update -c")
-
-	// Ensure global settings.json has required env variables
-	if err := ensureGlobalSettingsEnv(); err != nil {
-		_, _ = fmt.Fprintln(out, tui.CheckLine("warn", "Global settings env", "update failed", err.Error(), &th))
-	}
-
-	// Install pre-push hook (REQ-CIAUT-002). Non-fatal; --no-hooks opts out.
-	installPrePushHookOptional(projectRoot, getBoolFlag(cmd, "no-hooks"), out)
-
-	// Install pre-commit hook (REQ-PC-001). Fast-subset commit tier; --no-hooks opts out.
-	installPreCommitHookOptional(projectRoot, getBoolFlag(cmd, "no-hooks"), out)
-
-	// SPEC-WORKTREE-BRANCH-GUARD-001 (REQ-WBG-009): one-line worktree advisory on
-	// update completion. The primary checkout is shared; branch-changing work
-	// belongs in a worktree.
-	emitWorktreeAdvisory(out, projectRoot)
-
-	return nil
-}
-
-// @MX:NOTE: [AUTO] runTemplateSyncWithProgress — M4-S4d-2 DDD migration. Console reporter
-// wrapper. Only the key outputs are converted: tui.Pill (skip), tui.Section (analyzing), tui.Pill (cancel).
-//
-// @MX:NOTE: [AUTO] SPEC-V3R6-UPDATE-ARCHIVE-CONTRACT-001 — return shape extended
-// to (skipped bool, err error) so that runUpdate can short-circuit the
-// downstream legacy-skill archive block when sync is skipped (version-match
-// without --force). skipped == true means "no template files were written";
-// callers MUST NOT trigger archive checks in that case.
-//
-// runTemplateSyncWithProgress runs template sync with simple console output.
-// Return values:
-//   - skipped: true when version matches and --force is absent (sync was a no-op).
-//     When skipped == true, callers should not trigger downstream archive checks.
-//   - err: any non-skip error encountered.
-//
-// A skipped sync is not an error; callers receive (true, nil).
-// A user-cancelled merge is also (true, nil) — it is a no-op for downstream
-// purposes (no files written), so archive does not need to run.
-func runTemplateSyncWithProgress(cmd *cobra.Command) (skipped bool, err error) {
-	out := cmd.OutOrStdout()
-	th := resolveTheme()
-	projectRoot := "."
-	autoConfirm := getBoolFlag(cmd, "yes")
-	forceUpdate := getBoolFlag(cmd, "force")
-
-	// Use printer-backed console output for progress reporting
-	// (REQ-CTX-015: routes step events through the Printer to stderr).
-	consoleReporter := newPrinterReporter(
-		printer.New(printer.WithWriters(cmd.OutOrStdout(), cmd.ErrOrStderr())))
-
-	// Check for version match before proceeding
-	packageVersion := version.GetVersion()
-	projectVersion, verr := plan.GetProjectConfigVersion(projectRoot)
-	if verr == nil && packageVersion == projectVersion && !forceUpdate {
-		_, _ = fmt.Fprintln(out)
-		_, _ = fmt.Fprintln(out, tui.Pill(tui.PillOpts{Kind: tui.PillOk, Solid: false, Label: report.RenderOutcome(report.OutcomeAlreadyUpToDate, 0, ""), Theme: &th}))
-		return true, nil
-	}
-
-	// Confirm merge before proceeding (unless auto-confirm is set)
-	if !autoConfirm {
-		embedded, eerr := template.EmbeddedTemplates()
-		if eerr != nil {
-			return false, fmt.Errorf("load embedded templates: %w", eerr)
-		}
-
-		deployer := template.NewDeployerWithForceUpdate(embedded, true)
-		analysis := updatemerge.AnalyzeMergeChanges(deployer, projectRoot)
-
-		_, _ = fmt.Fprintln(out)
-		_, _ = fmt.Fprintln(out, tui.Section("Analyzing merge changes", tui.SectionOpts{Theme: &th}))
-		proceed, cerr := confirmViaPreview(analysis, projectRoot)
-		if cerr != nil {
-			return false, fmt.Errorf("confirm merge for %d files (risk: %s): %w",
-				len(analysis.Files), analysis.RiskLevel, cerr)
-		}
-		if !proceed {
-			_, _ = fmt.Fprintln(out)
-			_, _ = fmt.Fprintln(out, tui.Pill(tui.PillOpts{Kind: tui.PillNeutral, Solid: false, Label: "Merge cancelled by user", Theme: &th}))
-			return true, nil
-		}
-	}
-
-	return false, runTemplateSyncWithReporter(cmd, consoleReporter, true)
-}
-
-// toPreviewInputs maps a merge.MergeAnalysis into the neutral
-// update.FilePreviewInput slice consumed by the preview entry point (M3c
-// convergence, SPEC-CLI-TUX-V3-003 plan Known Issue #7). The Exists/Conflict
-// derivation mirrors the same signals the deploy stage enforces — Exists from
-// os.Stat (mirrors analyzeFiles), Conflict from RiskLevel=="high" (mirrors
-// buildMergeAnalysis hasConflicts). The classification itself (which file is
-// preserved/added/updated/conflict) is derived downstream by update.Classify
-// via the injected isUserOwnedNamespace predicate (REQ-TUX3-001/002 single
-// source of truth — this mapping introduces NO parallel heuristic).
-func toPreviewInputs(analysis merge.MergeAnalysis, projectRoot string) []update.FilePreviewInput {
-	inputs := make([]update.FilePreviewInput, 0, len(analysis.Files))
-	for _, fa := range analysis.Files {
-		_, statErr := os.Stat(filepath.Join(projectRoot, fa.Path))
-		inputs = append(inputs, update.FilePreviewInput{
-			RelPath:  fa.Path,
-			Exists:   statErr == nil,
-			Conflict: fa.RiskLevel == "high",
-			Diff:     fa.Changes,
-		})
-	}
-	return inputs
-}
-
-// confirmViaPreview is the single convergence surface for BOTH merge.ConfirmMerge
-// call sites in this file (plan Known Issue #7). It launches the Bubble Tea v2
-// TUI (AC-TUX3-008/009) when stdin is a terminal, classifying through
-// update.Classify with the shared isUserOwnedNamespace predicate
-// (REQ-TUX3-001/002/014 — no parallel heuristic, `preserved (user-owned)` label
-// derived from the shared predicate).
-//
-// Behavior preservation relative to the prior merge.ConfirmMerge calls: both
-// call sites reach this helper ONLY on the interactive (non-skipped, non-auto)
-// path — the skip-confirm and --yes (autoConfirm) branches short-circuit
-// upstream and set proceed=true directly, byte-identical to before. When stdin
-// is NOT a terminal, the user cannot be asked interactively, so this helper
-// returns an error directing them to --yes — mirroring both the explicit
-// Windows guard that merge.ConfirmMerge carried (confirm.go REQ-CFS-007/008)
-// and the de-facto tea.Run() non-TTY failure on darwin/linux that the v1
-// ConfirmMerge relied on. It MUST NOT silently proceed in the non-TTY case:
-// that would bypass the confirmation gate the caller explicitly entered (the
-// caller did not pass --yes), changing which files get deployed. The
-// preview-fallback's proceed=true semantics belong to the --yes abstraction,
-// which never reaches this helper.
-func confirmViaPreview(analysis merge.MergeAnalysis, projectRoot string) (bool, error) {
-	if !isatty.IsTerminal(os.Stdin.Fd()) {
-		return false, fmt.Errorf("merge confirmation UI requires an interactive terminal; " +
-			"rerun with --yes to auto-confirm in a non-TTY environment")
-	}
-	return update.PreviewClassification(toPreviewInputs(analysis, projectRoot), plan.IsUserOwnedNamespace, update.PreviewOptions{Interactive: true})
-}
-
 // @MX:NOTE: [AUTO] runShellEnvConfig — M4-S4d-2 DDD migration. tui.Section header,
 // tui.KV (Shell/Config/Explanation), tui.CheckLine (Changes list), tui.Pill (final outcome).
 //
@@ -1271,337 +702,6 @@ func runAgencyMigrationAdapter(projectRoot string, dryRun, force bool, out io.Wr
 		}
 		return fmt.Errorf("agency migration: %w", runErr)
 	}
-	return nil
-}
-
-// runInitWizard runs the configuration wizard for reconfiguring an existing project.
-// Used by 'moai update -c/--config' to edit project settings.
-// @MX:NOTE: [AUTO] runInitWizard — M4-S4d-3 DDD migration. tui.Pill (uninitialized warning,
-// cancelled, success) + tui.Section (Reconfiguration header; AC-017 emoji removed).
-func runInitWizard(cmd *cobra.Command, reconfigure bool) error {
-	out := cmd.OutOrStdout()
-	th := resolveTheme()
-
-	// Verify the project is initialized
-	cwd, err := os.Getwd()
-	if err != nil {
-		return fmt.Errorf("get working directory: %w", err)
-	}
-
-	if _, err := os.Stat(filepath.Join(cwd, defs.MoAIDir)); os.IsNotExist(err) {
-		_, _ = fmt.Fprintln(out, tui.Pill(tui.PillOpts{Kind: tui.PillWarn, Solid: false, Label: "Project not initialized · Run 'moai init' first", Theme: &th}))
-		return fmt.Errorf("project not initialized")
-	}
-
-	// Print banner and welcome message
-	uikit.PrintBanner(version.GetVersion())
-	if reconfigure {
-		_, _ = fmt.Fprintln(out, tui.Section("Project Reconfiguration Wizard", tui.SectionOpts{Theme: &th}))
-		_, _ = fmt.Fprintln(out, "This wizard will help you update your project configuration.")
-	} else {
-		uikit.PrintWelcomeMessage()
-	}
-
-	// REQ-1: Read locale from language.yaml
-	locale := wizard.ReadLocaleFromProject(cwd)
-
-	// REQ-2: Read existing username from config (used as default value)
-	existingGitHubUsername := wizard.ReadGitHubUsernameFromConfig(cwd)
-	existingGitLabUsername := wizard.ReadGitLabUsernameFromConfig(cwd)
-
-	// REQ-3: Check whether gh CLI is authenticated
-	ghAuthenticated := wizard.IsGhAuthenticated()
-
-	// Generate default questions and set defaults from existing values
-	questions := wizard.ReconfigureQuestions(cwd)
-	if existingGitHubUsername != "" {
-		if q := wizard.QuestionByID(questions, "github_username"); q != nil {
-			q.Default = existingGitHubUsername
-		}
-	}
-	if existingGitLabUsername != "" {
-		if q := wizard.QuestionByID(questions, "gitlab_username"); q != nil {
-			q.Default = existingGitLabUsername
-		}
-	}
-	// REQ-3: Skip github_token question when gh auth is authenticated
-	if ghAuthenticated {
-		if q := wizard.QuestionByID(questions, "github_token"); q != nil {
-			q.Condition = func(_ *wizard.WizardResult) bool { return false }
-		}
-	}
-
-	// Run wizard with locale and custom questions
-	result, err := wizard.RunWithLocale(questions, nil, locale)
-	if err != nil {
-		if errors.Is(err, wizard.ErrCancelled) {
-			_, _ = fmt.Fprintln(out, tui.Pill(tui.PillOpts{Kind: tui.PillNeutral, Solid: false, Label: "Configuration cancelled", Theme: &th}))
-			return nil
-		}
-		return fmt.Errorf("wizard failed: %w", err)
-	}
-
-	// Apply configuration updates to .moai/config/sections/
-	// This updates the YAML configuration files based on wizard results
-	if err := applyWizardConfig(cwd, result); err != nil {
-		return fmt.Errorf("apply configuration: %w", err)
-	}
-
-	// F1 security fix: a token entered in the wizard is NOT persisted to any
-	// git-tracked config. Tell the user so the credential is not silently
-	// dropped — direct them to the provider CLI credential store.
-	if result.GitHubToken != "" {
-		_, _ = fmt.Fprintln(out, tui.Pill(tui.PillOpts{Kind: tui.PillWarn, Solid: false, Label: "GitHub token was not saved (never stored in plaintext). Run 'gh auth login' to authenticate.", Theme: &th}))
-	}
-	if result.GitLabToken != "" {
-		_, _ = fmt.Fprintln(out, tui.Pill(tui.PillOpts{Kind: tui.PillWarn, Solid: false, Label: "GitLab token was not saved (never stored in plaintext). Run 'glab auth login' to authenticate.", Theme: &th}))
-	}
-
-	_, _ = fmt.Fprintln(out, tui.Pill(tui.PillOpts{Kind: tui.PillOk, Solid: false, Label: "Configuration updated successfully", Theme: &th}))
-
-	return nil
-}
-
-// applyWizardConfig applies wizard results to the project configuration files.
-func applyWizardConfig(projectRoot string, result *wizard.WizardResult) error {
-	// F3: validate externally-supplied identity fields before any write so a
-	// malformed URL or username never lands in a persisted config file.
-	if err := validateWizardInput(result); err != nil {
-		return err
-	}
-
-	sectionsDir := filepath.Join(projectRoot, defs.MoAIDir, defs.SectionsSubdir)
-
-	// user.yaml: Save GitHub/GitLab username plus the wizard-collected user
-	// display name. Access tokens are DELIBERATELY NOT persisted here (F1
-	// security fix): user.yaml is git-tracked and world-readable, so writing a
-	// secret into it leaks the credential. Credentials are delegated to the
-	// `gh` / `glab` CLI (see the advisory surfaced by runInitWizard). The block
-	// triggers on any non-secret user field (username or display name), so a
-	// reconfigure that answers only the user_name question still persists
-	// user.name — but a token-only answer creates nothing.
-	hasUserFields := result.GitHubUsername != "" ||
-		result.GitLabUsername != "" ||
-		result.UserName != ""
-	if hasUserFields {
-		userPath := filepath.Join(sectionsDir, defs.UserYAML)
-		// Read existing file
-		userData, err := os.ReadFile(userPath)
-		if err != nil && !os.IsNotExist(err) {
-			return fmt.Errorf("read user.yaml: %w", err)
-		}
-
-		// Parse YAML
-		var user map[string]any
-		if len(userData) > 0 {
-			if err := yaml.Unmarshal(userData, &user); err != nil {
-				return fmt.Errorf("parse user.yaml: %w", err)
-			}
-		} else {
-			user = make(map[string]any)
-		}
-
-		// Ensure user.user section exists
-		var userConfig map[string]any
-		if existingUser, ok := user["user"].(map[string]any); ok {
-			userConfig = existingUser
-		} else {
-			userConfig = make(map[string]any)
-		}
-
-		// Save GitHub/GitLab usernames (non-secret). Access tokens
-		// (result.GitHubToken / result.GitLabToken) are intentionally NOT
-		// written — persisting a secret into the git-tracked user.yaml is the
-		// F1 vulnerability this fix closes. Tokens are delegated to the
-		// `gh` / `glab` CLI credential store instead.
-		if result.GitHubUsername != "" {
-			userConfig["github_username"] = result.GitHubUsername
-		}
-		if result.GitLabUsername != "" {
-			userConfig["gitlab_username"] = result.GitLabUsername
-		}
-
-		// Save user display name (reconfigure wizard user_name question)
-		if result.UserName != "" {
-			userConfig["name"] = result.UserName
-		}
-
-		user["user"] = userConfig
-
-		// Save to file
-		updatedData, err := yaml.Marshal(user)
-		if err != nil {
-			return fmt.Errorf("marshal user.yaml: %w", err)
-		}
-		if err := os.WriteFile(userPath, updatedData, defs.FilePerm); err != nil {
-			return fmt.Errorf("write user.yaml: %w", err)
-		}
-	}
-
-	// language.yaml: Save conversation language (reconfigure wizard language
-	// question). Sets both conversation_language and conversation_language_name
-	// while preserving all sibling keys (agent_prompt_language, code_comments, ...).
-	if result.ConversationLang != "" {
-		langPath := filepath.Join(sectionsDir, defs.LanguageYAML)
-		langData, err := os.ReadFile(langPath)
-		if err != nil && !os.IsNotExist(err) {
-			return fmt.Errorf("read language.yaml: %w", err)
-		}
-
-		var lang map[string]any
-		if len(langData) > 0 {
-			if err := yaml.Unmarshal(langData, &lang); err != nil {
-				return fmt.Errorf("parse language.yaml: %w", err)
-			}
-		} else {
-			lang = make(map[string]any)
-		}
-
-		// Ensure language section exists
-		var language map[string]any
-		if existing, ok := lang["language"].(map[string]any); ok {
-			language = existing
-		} else {
-			language = make(map[string]any)
-		}
-
-		language["conversation_language"] = result.ConversationLang
-		language["conversation_language_name"] = result.ConversationLang
-		lang["language"] = language
-
-		updatedData, err := yaml.Marshal(lang)
-		if err != nil {
-			return fmt.Errorf("marshal language.yaml: %w", err)
-		}
-		if err := os.WriteFile(langPath, updatedData, defs.FilePerm); err != nil {
-			return fmt.Errorf("write language.yaml: %w", err)
-		}
-	}
-
-	// git-strategy.yaml: Save git mode and provider (REQ-4)
-	if result.GitMode != "" || result.GitProvider != "" {
-		gitStratPath := filepath.Join(sectionsDir, defs.GitStrategyYAML)
-		gsData, err := os.ReadFile(gitStratPath)
-		if err != nil && !os.IsNotExist(err) {
-			return fmt.Errorf("read git-strategy.yaml: %w", err)
-		}
-
-		var gs map[string]any
-		if len(gsData) > 0 {
-			if err := yaml.Unmarshal(gsData, &gs); err != nil {
-				return fmt.Errorf("parse git-strategy.yaml: %w", err)
-			}
-		} else {
-			gs = make(map[string]any)
-		}
-
-		// Ensure git_strategy section exists
-		var gitStrategy map[string]any
-		if existing, ok := gs["git_strategy"].(map[string]any); ok {
-			gitStrategy = existing
-		} else {
-			gitStrategy = make(map[string]any)
-		}
-
-		if result.GitMode != "" {
-			gitStrategy["mode"] = result.GitMode
-		}
-		if result.GitProvider != "" {
-			gitStrategy["provider"] = result.GitProvider
-		}
-
-		// Save GitLab instance URL (REQ-5)
-		if result.GitLabInstanceURL != "" {
-			var gitlabSection map[string]any
-			if existing, ok := gitStrategy["gitlab"].(map[string]any); ok {
-				gitlabSection = existing
-			} else {
-				gitlabSection = make(map[string]any)
-			}
-			gitlabSection["instance_url"] = result.GitLabInstanceURL
-			gitStrategy["gitlab"] = gitlabSection
-		}
-
-		gs["git_strategy"] = gitStrategy
-
-		updatedData, err := yaml.Marshal(gs)
-		if err != nil {
-			return fmt.Errorf("marshal git-strategy.yaml: %w", err)
-		}
-		if err := os.WriteFile(gitStratPath, updatedData, defs.FilePerm); err != nil {
-			return fmt.Errorf("write git-strategy.yaml: %w", err)
-		}
-	}
-
-	// SPEC-MODEL-PROFILE-MATRIX-001 (REQ-MPM-016/024): the former plan_type × tier
-	// agent-frontmatter mutation is RETIRED — persist the resolved model policy to
-	// llm.profile (normalized {high,medium,low} → {max,medium,low}) instead of
-	// mutating agent frontmatter, which stays at model: inherit.
-	if result.ModelPolicy != "" {
-		policy := template.ModelPolicy(result.ModelPolicy)
-		if template.IsValidModelPolicy(string(policy)) {
-			if err := template.ApplyProfile(projectRoot, template.NormalizeToTier(result.ModelPolicy)); err != nil {
-				return fmt.Errorf("apply profile: %w", err)
-			}
-			// Persist model_policy to system.yaml so it survives future updates
-			systemPath := filepath.Join(sectionsDir, defs.SystemYAML)
-			systemData, _ := os.ReadFile(systemPath)
-			var sys map[string]any
-			if len(systemData) > 0 {
-				_ = yaml.Unmarshal(systemData, &sys)
-			}
-			if sys == nil {
-				sys = make(map[string]any)
-			}
-			moaiSection, _ := sys["moai"].(map[string]any)
-			if moaiSection == nil {
-				moaiSection = make(map[string]any)
-			}
-			moaiSection["model_policy"] = string(policy)
-			sys["moai"] = moaiSection
-			if updatedData, err := yaml.Marshal(sys); err == nil {
-				_ = os.WriteFile(systemPath, updatedData, defs.FilePerm)
-			}
-		}
-	}
-
-	// quality.yaml: Save development mode (REQ-4)
-	if result.DevelopmentMode != "" {
-		qualityPath := filepath.Join(sectionsDir, defs.QualityYAML)
-		qualityData, err := os.ReadFile(qualityPath)
-		if err != nil && !os.IsNotExist(err) {
-			return fmt.Errorf("read quality.yaml: %w", err)
-		}
-
-		var quality map[string]any
-		if len(qualityData) > 0 {
-			if err := yaml.Unmarshal(qualityData, &quality); err != nil {
-				return fmt.Errorf("parse quality.yaml: %w", err)
-			}
-		} else {
-			quality = make(map[string]any)
-		}
-
-		// Ensure constitution section exists
-		var constitution map[string]any
-		if existing, ok := quality["constitution"].(map[string]any); ok {
-			constitution = existing
-		} else {
-			constitution = make(map[string]any)
-		}
-
-		constitution["development_mode"] = result.DevelopmentMode
-		quality["constitution"] = constitution
-
-		updatedData, err := yaml.Marshal(quality)
-		if err != nil {
-			return fmt.Errorf("marshal quality.yaml: %w", err)
-		}
-		if err := os.WriteFile(qualityPath, updatedData, defs.FilePerm); err != nil {
-			return fmt.Errorf("write quality.yaml: %w", err)
-		}
-	}
-
 	return nil
 }
